with error 10

+++ CALC/calculating.h
@@ -9,7 +9,7 @@ void RPN(stack **first, stack **polish_num, stack **polish_support);
 void calculator(stack ** result, stack **calculated);
 int QT_processing (char *input, long double value, long double *result);
 
-long double bank_(long double sum, int time,long double percent, size_t type, long double *sum_res, long double *mounth_pay, long double *overpay, int N);
+int bank_(long double sum, int time,long double percent, size_t type, long double *sum_res, long double *mounth_pay, long double *overpay, int N);
 int check_for_bank(long double sum, int time,long double percent, size_t type);
 
 #endif  // SMART_CALCULATOR_CALCULATING_H
